feat(gwas-ui-demo): add quantitative vs case control selection
