+++ src/Analysis/GWASUIApp/GWASUIApp.css
@@ -86,6 +86,15 @@
   font-style: italic;
 }
 
+.GWASUI-typeSelector {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  padding: 10px;
+}
+
 .ant-select-selector, .ant-select-selection-item, .ant-input, .ant-input-disabled {
   background-color: white;
   color: black;
